+++ tests/data/included_w/src/bar/Sub/EwokPack/tests/unit/BasicTest.php
@@ -1,0 +1,10 @@
+<?php
+namespace EwokPack\Test;
+
+class BasicTest extends \Codeception\TestCase\Test
+{
+    function testAssert()
+    {
+        $this->assertTrue(true);
+    }
+}